rust-fsp-wrapper: move fsp calls from rust-ipl.
TBD: Need refactor.
