+++ rust-fsp-wrapper/Cargo.toml
@@ -8,3 +8,6 @@ edition = "2018"
 
 [dependencies]
 scroll = { version = "0.10", default-features=false, features = ["derive"] }
+rust-firmware-layout = { path = "../rust-firmware-layout" }
+log = { path = "../rust-ipl-log", package="rust-ipl-log" }
+x86 = "0.34.0"
